Moved Python code to own module
Having the code in an own module facilitates development a lot.
Previously it was handled as a string within TypeScript code. That made
syntax highlighting and indentation really hard.
Now it is in a separate Python file that can be installed and imported
as a module.
Unfortunately this reduces access to the global namespace
which is the explicit goal of the module. This is currently solved by
importing __main__, but it might be a good idea to instead just pass the
globals and locals dict. I have not tried that yet.